+++ examples/freertos/src/FreeRTOSConfig.h
@@ -56,7 +56,8 @@ extern void vAssertCalled(const char *file, int line);
 #define configUSE_DAEMON_TASK_STARTUP_HOOK 0
 #define configCPU_CLOCK_HZ ((unsigned long)20000000)
 #define configTICK_RATE_HZ ((TickType_t)1000)
-#define configMINIMAL_STACK_SIZE ((unsigned short)2000)
+#define configMINIMAL_STACK_SIZE ((unsigned short)1000)
+#define configMINIMAL_SECURE_STACK_SIZE (1024)
 #define configTOTAL_HEAP_SIZE ((size_t)(2048))
 #define configMAX_TASK_NAME_LEN (10)
 #define configUSE_TRACE_FACILITY 1