+++ ports/zephyr/include/memfault/ports/zephyr/fota.h
@@ -12,6 +12,19 @@
 extern "C" {
 #endif
 
+//! @brief A callback invoked after a download has completed
+//!
+//! @note The default implementation will reboot the system after an OTA download has completed.
+//! A boolean return value is added to allow for custom implementations that may want to handle the
+//! download completion differently (e.g. in a testing framework context where returning from the
+//! FOTA will allow the result of the image download to be captured).
+//!
+//! A custom implementation can be provided by setting
+//! CONFIG_MEMFAULT_ZEPHYR_FOTA_DOWNLOAD_CALLBACK_CUSTOM=y and implementing this function
+//!
+//! @return true if the download was successful, false otherwise
+bool memfault_zephyr_fota_download_callback(void);
+
 //! @brief Initiates a FOTA update for Zephyr systems
 //!
 //! @note This function will block until the FOTA update is complete. If the update is successful,