+++ ports/zephyr/v2.4/memfault_fault_handler.c
@@ -34,6 +34,11 @@
   #define MEMFAULT_ZEPHYR_NMI_HANDLER_SET z_NmiHandlerSet
 extern void MEMFAULT_ZEPHYR_NMI_HANDLER_SET(void (*pHandler)(void));
 #endif  // MEMFAULT_ZEPHYR_VERSION_GT(3, 3)
+
+// This header is used on Zephyr 3.7+ to get the exception frame declaration
+#if MEMFAULT_ZEPHYR_VERSION_GT(3, 6)
+#include <zephyr/arch/exception.h>
+#endif
 // clang-format on
 
 // By default, the Zephyr NMI handler is an infinite loop. Instead
@@ -54,9 +59,15 @@ SYS_INIT(prv_install_nmi_handler, APPLICATION, CONFIG_KERNEL_INIT_PRIORITY_DEFAU
 // Note: There is no header exposed for this zephyr function
 extern void sys_arch_reboot(int type);
 
-// Intercept zephyr/kernel/fatal.c:z_fatal_error()
+// Intercept zephyr/kernel/fatal.c:z_fatal_error(). Note that the signature
+// changed in zephyr 3.7.
+#if MEMFAULT_ZEPHYR_VERSION_GT(3, 6)
+void __wrap_z_fatal_error(unsigned int reason, const struct arch_esf *esf);
+void __real_z_fatal_error(unsigned int reason, const struct arch_esf *esf);
+#else
 void __wrap_z_fatal_error(unsigned int reason, const z_arch_esf_t *esf);
 void __real_z_fatal_error(unsigned int reason, const z_arch_esf_t *esf);
+#endif
 // Ensure the substituted function signature matches the original function
 _Static_assert(__builtin_types_compatible_p(__typeof__(&z_fatal_error),
                                             __typeof__(&__wrap_z_fatal_error)) &&
@@ -64,7 +75,12 @@ _Static_assert(__builtin_types_compatible_p(__typeof__(&z_fatal_error),
                                               __typeof__(&__real_z_fatal_error)),
                "Error: check z_fatal_error function signature");
 
-void __wrap_z_fatal_error(unsigned int reason, const z_arch_esf_t *esf) {
+#if MEMFAULT_ZEPHYR_VERSION_GT(3, 6)
+void __wrap_z_fatal_error(unsigned int reason, const struct arch_esf *esf)
+#else
+void __wrap_z_fatal_error(unsigned int reason, const z_arch_esf_t *esf)
+#endif
+{
   // flush logs prior to capturing coredump & rebooting
   LOG_PANIC();